requirements.txt
google-adk
google-generativeai
streamlit